+++ scripts/run-github-action.sh
@@ -33,6 +33,7 @@ max_inventory_entries=${INPUT_MAX_INVENTORY_ENTRIES:-32768}
 max_portfolio_source_bytes=${INPUT_MAX_PORTFOLIO_SOURCE_BYTES:-4294967296}
 max_portfolio_snapshot_cells=${INPUT_MAX_PORTFOLIO_SNAPSHOT_CELLS:-2000000}
 max_dependency_edges=${INPUT_MAX_DEPENDENCY_EDGES:-2000000}
+max_formula_defined_name_states=${INPUT_MAX_FORMULA_DEFINED_NAME_STATES:-1000000}
 max_change_analysis_states=${INPUT_MAX_CHANGE_ANALYSIS_STATES:-100000}
 max_report_bytes=${INPUT_MAX_REPORT_BYTES:-33554432}
 max_link_impact=${INPUT_MAX_LINK_IMPACT:-100000}
@@ -130,6 +131,9 @@ if ! [[ "$max_portfolio_snapshot_cells" =~ ^[1-9][0-9]*$ ]]; then
 fi
 if ! [[ "$max_dependency_edges" =~ ^[1-9][0-9]*$ ]]; then
   fail 'max-dependency-edges must be a positive integer.'
+fi
+if ! [[ "$max_formula_defined_name_states" =~ ^[1-9][0-9]*$ ]]; then
+  fail 'max-formula-defined-name-states must be a positive integer.'
 fi
 if ! [[ "$max_change_analysis_states" =~ ^[1-9][0-9]*$ ]]; then
   fail 'max-change-analysis-states must be a positive integer.'
@@ -234,6 +238,7 @@ else
   command+=(diff "$baseline" "$candidate")
 fi
 command+=(--max-dependency-edges "$max_dependency_edges")
+command+=(--max-formula-defined-name-states "$max_formula_defined_name_states")
 command+=(--max-change-analysis-states "$max_change_analysis_states")
 command+=(--max-report-bytes "$max_report_bytes")
 command+=(--format "$format" --output "$report_path" --fail-on "$fail_on")